Reuse the same transport for every request
By making the transport for every request a global variable, we can
take advantage of the connection caching that transports provide by
default. This should reduce the overhead of making a request, since
the connection doesn't have to be reestablished. It also prevents
connections from leaking with new transports for every request.