Accommodate Eureka version 1.22's format for ports
In version 1.22, the Eureka server changed its JSON encoding for an
instance's port numbers: instead of writing the port numbers as JSON
string values, it started writing them as JSON number values. In order
to accommodate Eureka servers of versions on either side of this
change, accept instance port numbers as both strings and numbers, but
continue to emit them as strings, which Eureka servers of all versions
continue to accept.