Satisfy Eureka 1.2.2 and later when registering
As of Eureka version 1.2.2 and later, when registering an instance
encoded as JSON, one must supply a value for the

  instance.dataCenterInfo.@class

field, which must be set appropriately for the sibling "name" field's
value. Take care of setting this value for the "Amazon" and "MyOwn"
data center types, but allow callers to specify the value for their
own custom data center types for which we cannot anticipate a suitable
value.